Created Example Code for Triggers to MIDI
Four trigger inputs generate Note On messages for high signals, and Note Off messages (Note On messages with a velocity of 0) for low signals, using the Betweener Library. Note, trigger inputs are "active low" so a high gate signal will result in a "falling edge" on the trigger at the microcontroller.